+++ test/unit/spec/v2/view-builder/views/EnrollPollOktaVerifyView_spec.js
@@ -1,11 +1,11 @@
-import EnrollAuthenticatorOVPollView from 'v2/view-builder/views/ov/EnrollAuthenticatorOVPollView';
+import EnrollPollOktaVerifyView from 'v2/view-builder/views/ov/EnrollPollOktaVerifyView';
 import AppState from 'v2/models/AppState';
 import Settings from 'models/Settings';
 import $sandbox from 'sandbox';
 import BrowserFeatures from 'util/BrowserFeatures';
 import xhrAuthenticatorEnrollOktaVerifyQr from '../../../../../../playground/mocks/data/idp/idx/authenticator-enroll-ov-qr';
 
-describe('v2/view-builder/views/ov/EnrollAuthenticatorOVPollView', function () {
+describe('v2/view-builder/views/ov/EnrollPollOktaVerifyView', function () {
   beforeEach(function () {
     this.init = (currentAuthenticator = xhrAuthenticatorEnrollOktaVerifyQr.currentAuthenticator.value) => {
       const currentViewState = {
@@ -17,10 +17,10 @@ describe('v2/view-builder/views/ov/EnrollAuthenticatorOVPollView', function () {
       const appState = new AppState({
         currentAuthenticator
       });
-      spyOn(appState, 'hasRemediationObject').and.callFake((formName) => formName === 'enroll-with-another-channel');
+      spyOn(appState, 'hasRemediationObject').and.callFake((formName) => formName === 'select-enrollment-channel');
       spyOn(appState, 'trigger');
       const settings = new Settings({ baseUrl: 'http://localhost:3000' });
-      this.view = new EnrollAuthenticatorOVPollView({
+      this.view = new EnrollPollOktaVerifyView({
         el: $sandbox,
         appState,
         settings,
@@ -40,7 +40,7 @@ describe('v2/view-builder/views/ov/EnrollAuthenticatorOVPollView', function () {
     this.init();
     expect(BrowserFeatures.isAndroid).toHaveBeenCalled();
     expect(BrowserFeatures.isIOS).toHaveBeenCalled();
-    expect(this.view.options.appState.trigger).toHaveBeenCalledWith('switchForm', 'enroll-with-another-channel');
+    expect(this.view.options.appState.trigger).toHaveBeenCalledWith('switchForm', 'select-enrollment-channel');
   });
 
   it('triggers switchForm on appState when on android device to select channel', function () {
@@ -48,7 +48,7 @@ describe('v2/view-builder/views/ov/EnrollAuthenticatorOVPollView', function () {
     spyOn(BrowserFeatures, 'isAndroid').and.callFake(() => true);
     this.init();
     expect(BrowserFeatures.isAndroid).toHaveBeenCalled();
-    expect(this.view.options.appState.trigger).toHaveBeenCalledWith('switchForm', 'enroll-with-another-channel');
+    expect(this.view.options.appState.trigger).toHaveBeenCalledWith('switchForm', 'select-enrollment-channel');
   });
 
   it('renders QR code view when on desktop', function () {
